Move DSi code to DSi sections
