+++ source/arm9/dldi/sd.twl.c
@@ -4,96 +4,92 @@
 #include <nds/system.h>
 #include <nds/arm9/cache.h>
 
-#include <nds/arm9/nand.h>
-
 //---------------------------------------------------------------------------------
-bool nand_Startup() {
+bool sdio_Startup() {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
+	fifoSendValue32(FIFO_SDMMC,SDMMC_HAVE_SD);
+	while(!fifoCheckValue32(FIFO_SDMMC));
+	int result = fifoGetValue32(FIFO_SDMMC);
 
-	return true;
+	if(result==0) return false;
+
+	fifoSendValue32(FIFO_SDMMC,SDMMC_SD_START);
+
+	fifoWaitValue32(FIFO_SDMMC);
+
+	result = fifoGetValue32(FIFO_SDMMC);
+	
+	return result == 0;
 }
 
 //---------------------------------------------------------------------------------
-bool nand_IsInserted() {
+bool sdio_IsInserted() {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
+	fifoSendValue32(FIFO_SDMMC,SDMMC_SD_IS_INSERTED);
 
-	return true;
+	fifoWaitValue32(FIFO_SDMMC);
+
+	int result = fifoGetValue32(FIFO_SDMMC);
+
+	return result == 1;
 }
 
 //---------------------------------------------------------------------------------
-bool nand_ReadSectors(sec_t sector, sec_t numSectors,void* buffer) {
+bool sdio_ReadSectors(sec_t sector, sec_t numSectors,void* buffer) {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
 	FifoMessage msg;
 
 	DC_FlushRange(buffer,numSectors * 512);
 
-	msg.type = SDMMC_NAND_READ_SECTORS;
+	msg.type = SDMMC_SD_READ_SECTORS;
 	msg.sdParams.startsector = sector;
 	msg.sdParams.numsectors = numSectors;
 	msg.sdParams.buffer = buffer;
-
+	
 	fifoSendDatamsg(FIFO_SDMMC, sizeof(msg), (u8*)&msg);
 
 	fifoWaitValue32(FIFO_SDMMC);
 
 	int result = fifoGetValue32(FIFO_SDMMC);
-
+	
 	return result == 0;
 }
 
 //---------------------------------------------------------------------------------
-bool nand_WriteSectors(sec_t sector, sec_t numSectors,const void* buffer) {
+bool sdio_WriteSectors(sec_t sector, sec_t numSectors,const void* buffer) {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
 	FifoMessage msg;
 
 	DC_FlushRange(buffer,numSectors * 512);
 
-	msg.type = SDMMC_NAND_WRITE_SECTORS;
+	msg.type = SDMMC_SD_WRITE_SECTORS;
 	msg.sdParams.startsector = sector;
 	msg.sdParams.numsectors = numSectors;
 	msg.sdParams.buffer = (void*)buffer;
-
+	
 	fifoSendDatamsg(FIFO_SDMMC, sizeof(msg), (u8*)&msg);
 
 	fifoWaitValue32(FIFO_SDMMC);
 
 	int result = fifoGetValue32(FIFO_SDMMC);
-
+	
 	return result == 0;
 }
 
 
 //---------------------------------------------------------------------------------
-bool nand_ClearStatus() {
+bool sdio_ClearStatus() {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
 	return true;
 }
 
 //---------------------------------------------------------------------------------
-bool nand_Shutdown() {
+bool sdio_Shutdown() {
 //---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return false;
 	return true;
 }
 
-//---------------------------------------------------------------------------------
-ssize_t nand_GetSize() {
-//---------------------------------------------------------------------------------
-	if (!REG_DSIMODE) return 0;
-
-	fifoSendValue32(FIFO_SDMMC, SDMMC_NAND_SIZE);
-
-	fifoWaitValue32(FIFO_SDMMC);
-
-	return fifoGetValue32(FIFO_SDMMC);
-
-}
-/*const DISC_INTERFACE __io_dsisd = {
+const DISC_INTERFACE __io_dsisd = {
 	DEVICE_TYPE_DSI_SD,
 	FEATURE_MEDIUM_CANREAD | FEATURE_MEDIUM_CANWRITE,
 	(FN_MEDIUM_STARTUP)&sdio_Startup,
@@ -103,4 +99,3 @@ ssize_t nand_GetSize() {
 	(FN_MEDIUM_CLEARSTATUS)&sdio_ClearStatus,
 	(FN_MEDIUM_SHUTDOWN)&sdio_Shutdown
 };
-*/
